fixed-size_data: Defining structure types Structure are not embedded in the programming language, it is up to the programmer to define them using structure type definitions. In DrRacket => (define-struct StructureName [FieldName ...])
A structure type definition defines several values simultaneously:
- one constructor (a function that creates structure instances);
- a selector per field which extracts the value of a given field
from an instance;
- a structure predicate which distinguishes instances from other
kinds of variables;

Structure is short for structure instance.
The name structure type is a generic name for the collection of
all possible instances.

Nesting information is natural and the best way to represent this
is by using structure instances.