fixed-size_data: Add cat-and-cham mix up program
Although the program is working as expected, I have not been following
the design recipe, hoping to speed of the exercises and finish the first
chapter.
Both these exercises were easy to implement as they did not produce many
errors after the first run. I am aware that this is due to the fact that
the original VCat and VCham programs, of which these functions are based,
had followed the design recipe and implemented the required tests.