fixed-size_data: Final exercise (yay!)
A checked version of a equality predicate guarantees that the data is solely inside a given collection.
Note: from now on, use the key=? equality predicate when handling key events and the mouse=? equality predicate when handling mouse events. Although all key and mouse events are strings, not all strings are key or mouse events.
These equality predicates will signal an error and thus inform the developer/user that something is wrong