creating_lists: Add first definitions of BSL lists
char, boolean and numbers are atomic values. They have the size of 1 unit.
Arbitrarily large data is data of finite yet arbitrary size.
List are such arbitrarily large data.
In BSL, an empty list is represented with '() and to add something to it
one must use the cons expressions.
In BSL, each list element might be of different type.