sicp: Add greatest common divisor
The GCD is the larger integer number which can be used to divide
2 integer numbers, resulting in a remainder of 0.
e.g. The greatest common divisor of 28 and 16 is 4

The GCD can be calculated, with a logarithm number of steps, by
assuming that the GCD of a and b is the same as b and r, being r
the remainder from dividing a with b.