sicp: Testing for primality
Searching for divisors
check all numbers from on to sqrt(n). If there is no divisor in
this interval, then then number is prime. The number of steps this
process requires grows in an order of sqrt(n)

Fermat’s little theory
If n is prime, every integer number lower than n, raised to the
n^th power, will be congruent with a modulo n. The number of steps
this algorithm requires is of order of growth log(n).
Here the answer is only probably correct.
    a modulo n => (remainder a n)
    congruent => two number are congruent modulo if both have the
same remainder when divided by n.
e.g. n = 17, a = 2
a ^ n = 2 ^ 17 = 131072
131072 % 17 = 2
2 == 2
=> 17 is likely a prime number