htdp: Design by composition - 191:192
Discussed the generalization of function. Using this design method
we get a simple main() function while the more complex auxiliary
functions do all the work.
It starts we a specific problem definition (i.e. draw a polygon)
and from there it is simple to generalize first (i.e. lets connect
some dots) and then used the auxiliary functions to reach the final
goal.