htdp: Design by composition - 193:194
Continue generalizing functions.
In these exercises we can see how functions are flexible, it is possible
to design the same function in several different ways (e.g. adding an
item to the begging of the list, adding an item to the end of the list or
by adding an extra input argument to simplify any other changes)
Although we were designing function that are already available in the
2htdp/image library, this helps us study the systematic way of designing
programs and gives us the first idea on why abstractions are important