htdp: Project: Lists - 196:198
Played around with dictionaries and composition.
Built a lot of auxiliary functions to get to the requested "main"
function.
Built function to:
- Verify a 1String is a Letter;
- Verify if a String starts with a given Letter;
- Count the number of words in a Dictionary that start with a Letter;
- Verify with first letter is most frequent (using max or sorted lists);
- Split a dictionary by letter;

For each function that should generate a list it was required to
have as input an empty list.
For each function with a counter, the counter had to be initialized
to 0 at the input of a function